UI: make our shortcut working from all appropriate places
